chore: update commit message pattern
To make the commit message more intuitive and scoped, the pattern has
been updated to mainly use the available applications and libs as the
prefix (e.g. portfolio: update footer page). This makes the commit
clearer on what parts of the workspace are affected.

An additional prefix "chore" is used for some generic chore-related
commits such as applying lints, config changes, etc.